requirements.txt
aiohttp
aiohttp-retry
aiosignal
asgiref
attrs
certifi
charset-normalizer
coreapi
coreschema
dj-database-url
Django
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
drf-yasg
frozenlist
gunicorn
idna
inflection
itypes
Jinja2
MarkupSafe
multidict
openapi-codec
packaging
phonenumbers
pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
PyYAML
requests
setuptools
simplejson
sqlparse
twilio
typing_extensions
uritemplate
urllib3
whitenoise
yarl
